Change to Apache License
After a narrow survey of the Java ecosystem, it looks like the Apache License is common. Indeed, some other code upon which we'll certainly depend (and likely copy from) is Apache licensed. Use this instead, as there are compatibility concerns with GPLv2; and OpenJDK, while being licensed GPLv2, has a linking exception.